+++ src/irisnet/noncore/legacy/srvresolver.cpp
@@ -159,13 +159,13 @@ void SrvResolver::stop()
 bool SrvResolver::isBusy() const
 {
 #ifndef NO_NDNS
-    if (d->nndns_busy || d->ndns.isBusy())
+ if (d->nndns_busy || d->ndns.isBusy())
 #else
-    if (d->nndns_busy)
-#endif
-        return true;
-    else
-        return false;
+ if (d->nndns_busy)
+ #endif
+    return true;
+ else
+    return false;
 }
 
 QList<Q3Dns::Server> SrvResolver::servers() const { return d->servers; }
@@ -198,8 +198,9 @@ void SrvResolver::nndns_resultsReady(const QList<XMPP::NameRecord> &results)
         // grab the server list and destroy the qdns object
         QList<Q3Dns::Server> list;
         for (int n = 0; n < results.count(); ++n) {
-            list += Q3Dns::Server(QString::fromLatin1(results[n].name()), results[n].priority(), results[n].weight(),
-                                  results[n].port());
+            list += Q3Dns::Server(QString::fromLatin1(results[n].name()), quint16(results[n].priority()),
+                                  quint16(results[n].weight()),
+                                  quint16(results[n].port()));
         }
 
         d->nndns_busy = false;
@@ -237,7 +238,7 @@ void SrvResolver::nndns_resultsReady(const QList<XMPP::NameRecord> &results)
             d->aaaa = true;
 
             d->resultAddress = list.first();
-            d->resultPort    = port;
+            d->resultPort    = quint16(port);
             resultsReady();
         } else {
             if (!d->aaaa)
@@ -268,7 +269,7 @@ void SrvResolver::ndns_done()
 
     if (!r.isNull()) {
         d->resultAddress = d->ndns.result();
-        d->resultPort    = port;
+        d->resultPort    = quint16(port);
         resultsReady();
     } else {
         // failed?  bail if last one
